Ammended NAN use to use 0 instead, and removed kalman filter

+++ base_firmware/lib/least_squares.h
@@ -4,7 +4,7 @@
 #define N_BEACONS 13
 #define N_AXIS    3
 
-int localise(float beacon_coords[N_BEACONS][N_AXIS], float rssi[N_BEACONS], float measured_power,
+int localise(float beacon_coords[N_BEACONS][N_AXIS], int8_t rssi[N_BEACONS], float measured_power,
              float path_loss_exp, float pos[N_AXIS]);
 
 #endif /* CENTRAL_H */
